feat(frontend): implement max_chars validation per message and keep files attached
- Add max_chars validation in submitPrompt before message submission
- Validate only current message files, not chat history
- Block message submission with toast error when limit is exceeded
- Remove file exclusion logic when character limit is exceeded
- Files remain attached even when exceeding limit, similar to max_count behavior
- Update validateFilesTotal API to accept optional chatId parameter
- Apply same behavior to both chat and channel MessageInput components
