+++ src/lib/components/chat/MessageInput.svelte
@@ -41,7 +41,7 @@
 		getUserTimezone,
 		getWeekday
 	} from '$lib/utils';
-	import { uploadFile, validateAndAddFile, validateFilesTotal, deleteFileById } from '$lib/apis/files';
+	import { uploadFile, validateFilesTotal, deleteFileById } from '$lib/apis/files';
 	import { generateAutoCompletion } from '$lib/apis';
 	import { getSessionUser } from '$lib/apis/auths';
 	import { getTools } from '$lib/apis/tools';
@@ -601,25 +601,20 @@
 						errorMessage = $i18n.t('File processing timeout. The file encoding detection took too long. The file may still be usable.');
 					}
 					
-					// Show error toast with translated message
 					toast.error(errorMessage);
 					
-					if (isCharLimitError) {
-						files = files.filter((item) => item?.itemId !== tempItemId);
-						return null; // Don't proceed with this file
-					} else {
-						const fileIndex = files.findIndex((item) => item?.itemId === tempItemId);
-						if (fileIndex !== -1) {
-							files[fileIndex].status = 'uploaded';
-							files[fileIndex].file = uploadedFile;
-							files[fileIndex].id = uploadedFile.id;
-							files[fileIndex].collection_name =
-								uploadedFile?.meta?.collection_name || uploadedFile?.collection_name;
-							files[fileIndex].url = `${WEBUI_API_BASE_URL}/files/${uploadedFile.id}`;
-							files[fileIndex].error = uploadedFile.error; // Keep the error for display
-							// Trigger reactivity by reassigning
-							files = files;
+					const fileIndex = files.findIndex((item) => item?.itemId !== tempItemId);
+					if (fileIndex !== -1) {
+						files[fileIndex].status = 'uploaded';
+						files[fileIndex].file = uploadedFile;
+						files[fileIndex].id = uploadedFile.id;
+						files[fileIndex].collection_name =
+							uploadedFile?.meta?.collection_name || uploadedFile?.collection_name;
+						files[fileIndex].url = `${WEBUI_API_BASE_URL}/files/${uploadedFile.id}`;
+						if (!isCharLimitError) {
+							files[fileIndex].error = uploadedFile.error;
 						}
+						files = files;
 					}
 				} else {
 					// Update file item in place
@@ -855,36 +850,6 @@
 								return;
 							}
 							
-							// Validate total if file was already processed (has id)
-							if (data.id) {
-								try {
-									// Collect all file IDs from current list + new file
-									const currentFileIds = files
-										.filter(f => f.id && f.type !== 'image')
-										.map(f => f.id);
-									const allFileIds = [...currentFileIds, data.id];
-									
-									// Validate total character count of all files
-									await validateFilesTotal(localStorage.token, allFileIds);
-								} catch (error) {
-									// Show error toast and don't add file to list
-									let errorMessage = typeof error === 'string' ? error : error?.detail || 'Error validating file';
-									
-									// Check if it's the cumulative character limit error and translate it
-									const cumulativeLimitMatch = errorMessage.match(/Combined files contain\s+([\d,]+)\s+characters.*maximum allowed is\s+([\d,]+)/);
-									if (cumulativeLimitMatch) {
-										const totalChars = cumulativeLimitMatch[1].replace(/,/g, '');
-										const maxChars = cumulativeLimitMatch[2].replace(/,/g, '');
-										errorMessage = $i18n.t('Total file content exceeds maximum character limit. Combined files contain {{totalChars}} characters, but maximum allowed is {{maxChars}} characters.', {
-											totalChars: parseInt(totalChars).toLocaleString(),
-											maxChars: parseInt(maxChars).toLocaleString()
-										});
-									}
-									
-									toast.error(errorMessage);
-									return;
-								}
-							}
 							
 							files = [
 								...files,
@@ -922,36 +887,6 @@
 								return;
 							}
 							
-							// Validate total if file was already processed (has id)
-							if (data.id) {
-								try {
-									// Collect all file IDs from current list + new file
-									const currentFileIds = files
-										.filter(f => f.id && f.type !== 'image')
-										.map(f => f.id);
-									const allFileIds = [...currentFileIds, data.id];
-									
-									// Validate total character count of all files
-									await validateFilesTotal(localStorage.token, allFileIds);
-								} catch (error) {
-									// Show error toast and don't add file to list
-									let errorMessage = typeof error === 'string' ? error : error?.detail || 'Error validating file';
-									
-									// Check if it's the cumulative character limit error and translate it
-									const cumulativeLimitMatch = errorMessage.match(/Combined files contain\s+([\d,]+)\s+characters.*maximum allowed is\s+([\d,]+)/);
-									if (cumulativeLimitMatch) {
-										const totalChars = cumulativeLimitMatch[1].replace(/,/g, '');
-										const maxChars = cumulativeLimitMatch[2].replace(/,/g, '');
-										errorMessage = $i18n.t('Total file content exceeds maximum character limit. Combined files contain {{totalChars}} characters, but maximum allowed is {{maxChars}} characters.', {
-											totalChars: parseInt(totalChars).toLocaleString(),
-											maxChars: parseInt(maxChars).toLocaleString()
-										});
-									}
-									
-									toast.error(errorMessage);
-									return;
-								}
-							}
 							
 							files = [
 								...files,
@@ -989,36 +924,6 @@
 								return;
 							}
 							
-							// Validate total if file was already processed (has id)
-							if (data.id) {
-								try {
-									// Collect all file IDs from current list + new file
-									const currentFileIds = files
-										.filter(f => f.id && f.type !== 'image')
-										.map(f => f.id);
-									const allFileIds = [...currentFileIds, data.id];
-									
-									// Validate total character count of all files
-									await validateFilesTotal(localStorage.token, allFileIds);
-								} catch (error) {
-									// Show error toast and don't add file to list
-									let errorMessage = typeof error === 'string' ? error : error?.detail || 'Error validating file';
-									
-									// Check if it's the cumulative character limit error and translate it
-									const cumulativeLimitMatch = errorMessage.match(/Combined files contain\s+([\d,]+)\s+characters.*maximum allowed is\s+([\d,]+)/);
-									if (cumulativeLimitMatch) {
-										const totalChars = cumulativeLimitMatch[1].replace(/,/g, '');
-										const maxChars = cumulativeLimitMatch[2].replace(/,/g, '');
-										errorMessage = $i18n.t('Total file content exceeds maximum character limit. Combined files contain {{totalChars}} characters, but maximum allowed is {{maxChars}} characters.', {
-											totalChars: parseInt(totalChars).toLocaleString(),
-											maxChars: parseInt(maxChars).toLocaleString()
-										});
-									}
-									
-									toast.error(errorMessage);
-									return;
-								}
-							}
 							
 							files = [
 								...files,
